fix: delete unnecessary files
