+++ nextjs/src/components/SharedHeader.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 
 import { useQuery } from '@tanstack/react-query';
 import { getFolderQueryKey, getUserQueryKey } from '@/api/queryKeys';
-import { getFolderApi, getSignedUserApi } from '@/api/apiCollection';
+import { getFolderApi, getSignedUserApi } from '@/api/api';
 
 import styled from 'styled-components';
