+++ nextjs/src/components/SigninForm.tsx
@@ -9,7 +9,7 @@ import { SignForm } from '@/styles/SignForm';
 
 import { SIGN_ERROR_MESSAGE } from '@/stores/constants';
 import { useMutation } from '@tanstack/react-query';
-import { postSigninApi } from '@/api/apiCollection';
+import { postSigninApi } from '@/api/api';
 import { EnteredSignInfo } from '@/types/apiType';
 import { saveAccessToken, saveRefreshToken } from '@/utils/manageTokenInfo';
