+++ nextjs/src/components/SignupForm.tsx
@@ -8,7 +8,7 @@ import { SignForm } from '@/styles/SignForm';
 
 import { SIGN_ERROR_MESSAGE } from '@/stores/constants';
 import { useMutation } from '@tanstack/react-query';
-import { postCheckEmailDuplicateApi, postSignupApi } from '@/api/apiCollection';
+import { postCheckEmailDuplicateApi, postSignupApi } from '@/api/api';
 import { EnteredEmail, EnteredSignInfo } from '@/types/apiType';
 import { useEffect, useState } from 'react';
 import { saveAccessToken, saveRefreshToken } from '@/utils/manageTokenInfo';
